requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.1.3
fastapi==0.102.0
gotrue==1.0.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
packaging==23.1
postgrest==0.10.8
pydantic==2.3.0
pydantic-core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
realtime==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
typing-extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
websockets==10.4
zipp==3.16.2
